extract blob components into their own folder
