+++ src/components/Blobs/Portrait.svelte
@@ -20,7 +20,7 @@
 </div>
 
 <style lang="scss">
-  @use '../styles/base';
+  @use '../../styles/base';
   
   .container {
     position: relative;
