+++ src/routes/__layout.svelte
@@ -2,8 +2,8 @@
   import Footer from "../components/Footer.svelte";
   import NavBar from "../components/NavBar.svelte";
   import Seo from "../components/Seo.svelte";
-  import Blob from "../components/Blob.svelte";
-  import BlobBackground from "../components/BlobBackground.svelte";
+  import Blob from "../components/Blobs/Blob.svelte";
+  import BlobBackground from "../components/Blobs/BlobBackground.svelte";
 </script>
 
 <svelte:head>
